% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daggnn-package.R
\docType{package}
\name{daggnn-package}
\alias{daggnn-package}
\alias{daggnn}
\title{daggnn}
\description{
daggnn implements a variational autoencoder to learn the DAG
 of a structural equations model
}
\references{
Yu, Yue and Chen, Jie and Gao, Tian and Yu, Mo (2019). DAG-GNN: DAG Structure
 Learning with Graph Neural Networks. ICML
}
\author{
Simon Dirmeier
}
