% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getters.R
\docType{methods}
\name{dataSet}
\alias{dataSet}
\alias{dataSet,sem_vae-method}
\title{Get the data set the SEM-VAE has been trained with}
\usage{
dataSet(obj)

\S4method{dataSet}{sem_vae}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the data}
}
\value{
returns a \code{matrix}
}
\description{
Returns the data set the SEM-VAE has been trained with
 as \code{matrix}
}
