% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot.sem_vae}
\alias{plot.sem_vae}
\title{Plot a simulated data set}
\usage{
\method{plot}{sem_vae}(x, ...)
}
\arguments{
\item{x}{a \code{sem_vae} object}

\item{...}{additional arguments passed to \code{ggplot2}}
}
\description{
Plot a simulated data set
}
