% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getters.R
\docType{methods}
\name{sem}
\alias{sem}
\alias{sem,sem_vae-method}
\title{Get the SEM form a SEM-VAE}
\usage{
sem(obj)

\S4method{sem}{sem_vae}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the SEM}
}
\value{
returns a \code{matrix}
}
\description{
Returns the SEM that underlies an S4 wrapper class
 as \code{matrix}
}
