% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{sem.vae}
\alias{sem.vae}
\title{Create a \code{sem.vae} object}
\usage{
sem.vae()
}
\value{
returns a \code{sem.vae} object
}
\description{
Create a new \code{sem.vae} object, i.e., a structural
 equations model variational autoencoder. The \code{sem.vae} can be
 used to infer the directed acyclic graph unerlying the SEM.
}
