% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{sem_vae-class}
\alias{sem_vae-class}
\title{SEM-VAE class}
\description{
A class representing a structural equations model variational autoencoder,
 i.e., a structural equations model variational autoencoder.
 The \code{sem_vae} can be used to infer the directed acyclic graph
 underlying the SEM.
}
\section{Slots}{

\describe{
\item{\code{sem}}{a \code{kerastools.model.RModel} object that encodes the
\code{sem_vae}}

\item{\code{data}}{the data set the SEM-VAE has been trained with}
}}

\references{
Yu, Yue and Chen, Jie and Gao, Tian and Yu, Mo (2019). DAG-GNN: DAG Structure
 Learning with Graph Neural Networks. ICML
}
